refactor(voyage): les seize gestes du compagnon sortent d'app.js
Démarrer un voyage, avancer d'une étape, ajouter ou retirer un arrêt, déplier
une jambe, ajuster ses SCU, la remettre à l'optimal — seize fonctions qui
rejoignent `voyage-gestes.js`.

La frontière du parcours devient nette, en trois modules :
- `voyage-donnees.ts` porte la DONNÉE et s'interdit de toucher au rendu ;
- `voyage-actions.ts` porte les deux gestes dont les VUES ont besoin
  (`pickJourney`, `syncViewsToJourney`) et la génération de la carte ;
- `voyage-gestes.js` porte les seize autres, appelés seulement par des
  délégations. Un fichier à part plutôt qu'un ajout au précédent : les vues
  importent l'un, jamais l'autre.

`toggleLegEditor` gagne un `flushSync` explicite. Il rendait le focus à
l'en-tête qu'on vient d'activer APRÈS un `renderJourney()` synchrone par
nature ; sous l'arbre, différé, le `?.` court-circuite et le focus retombe sur
`<body>`. À la souris ça ne se voit pas — au clavier on perd sa place, la
deuxième Entrée ne part plus de nulle part et Tab reprend au début du document.

QUATRE-VINGT-DIX-HUIT IMPORTS MORTS partent avec eux. Ce n'est pas du ménage :
c'est la mesure de ce que les huit migrations ont réellement déplacé. `app.js`
importait encore `commoditySummaries`, `multiTrips`, `offloadPlan`,
`tourneesEcoulement`, les dix cartes de `vues/` — tout le calcul de huit vues
qu'il ne rend plus. Sa liste d'imports passe de 26 à 23 lignes, mais surtout
elle ne ment plus sur ce qu'il fait.

app.js : 1 834 → 1 670 lignes.
Vérification : 244 e2e — 242 passés, 2 ignorés, 0 échec — · 490/490 unitaires ·
tsc silencieux · build OK.
